Refactor rex to separate out client library, and PM (process manager) library, so that the scheduler doesn't start the PM thread
