+++ riak_explorer/client/riak_explorer_client.go
